requirements.txt
fastapi
uvicorn
transformers
python-docx
pymupdf
python-multipart
torch